refactor: introduce Lombok annotations and update serializer references

+++ relay/Protocol/bedrock-codec/src/main/java/org/cloudburstmc/protocol/bedrock/codec/v898/serializer/AvailableCommandsSerializer_v898.java
@@ -38,7 +38,7 @@ public class AvailableCommandsSerializer_v898 extends AvailableCommandsSerialize
         helper.writeArray(buffer, commandData.getSubcommands(), (buf, subcommand) -> {
             int index = subCommands.indexOf(subcommand);
             checkArgument(index > -1, "Invalid subcommand index: " + subcommand);
-            buf.writeShortLE(index);
+            buf.writeIntLE(index);
         });
 
         CommandOverloadData[] overloads = commandData.getOverloads();
